Created custom hook for filter reducer
