+++ public/index.html
@@ -30,6 +30,7 @@
       rel="stylesheet"
       href="https://cosmic-ui.netlify.app/Components/utility.css"
     />
+    <link rel="stylesheet" href="../src/index.css" />
     <title>React App</title>
   </head>
   <body>
