+++ src/components/Filter/filter.css
@@ -37,6 +37,7 @@
   border-bottom: 1px solid var(--custom-pri-clr);
 }
 
-.products_filter_icon_resize {
+.products_filter_staricon_style {
   font-size: 0.8rem;
+  color: var(--dark-primary-clr);
 }
