+++ src/index.css
@@ -1,0 +1,8 @@
+:root {
+  --plant-pri-clr: #557669;
+  --plant-pri-light-clr: #94b49f;
+
+  --plant-sec-clr: #556052;
+
+  --black-clr: black;
+}
